move enum utils to verbs/ folder
