+++ javascript/react/src/components/verbs/renderers/forms/factories.ts
@@ -7,12 +7,12 @@ import { toggleListItem } from '@datashaper/utilities'
 import type { Step } from '@datashaper/workflow'
 import type { IDropdownOption } from '@fluentui/react'
 
-import { getEnumDropdownOptions } from '../../../../enums.js'
 import {
 	getLeftColumn,
 	getRightColumn,
 	getSimpleDropdownOptions,
 } from '../../../../hooks/index.js'
+import { getEnumDropdownOptions } from '../../enums.js'
 import type { CheckboxFormInput } from './CheckboxInput.types.js'
 import type { SingleChoiceFormInput } from './SingleChoiceInput.types.js'
 import type { FormInput } from './types.js'
